// types/scene.ts
export type Layer3Source =
  | 'sphere'
  | 'center'
  | 'ring'
  | 'disc'
  | 'cube'
  | 'cylinder'
  | 'random'
  | 'cone'
  | 'torus'
  | 'spiral'
  | 'galaxy'
  | 'grid'
  | 'plane'
  | 'image'
  | 'video'
  | 'text';